Move IncomeValidator interface to income.go
